Rename InputFormat according to the Hadoop package name.
